Refactor app-profile ui lifecycle redux flow into their own files
