+++ src/js/store.js
@@ -6,10 +6,8 @@ import startListeningForCurrentUser from './actions/current-user';
 import { startListeningForUserProfileChanges } from './actions/profile';
 import { startListeningForUserProfileImageChanges } from './actions/user-profile-image';
 import startListeningForAppUsers from './actions/app-users';
-import {
-  startListeningForProfileSocialMediaLinkChanges,
-  startListeningForProfileChanges
-} from './actions/profile-social-media-links';
+import { startListeningForProfileSocialMediaLinkChanges } from './actions/profile-social-media-links';
+import startListeningForProfileChanges from './actions/app-profiles';
 
 // const middleware = [thunk];
 // const enhancers = [];
